requirements.txt
langchain
langchain-openai
python-dotenv
moviepy
numpy
pillow
requests
